refactor: move flow service to operator
