+++ src/frontend/Cargo.toml
@@ -26,6 +26,7 @@ common-config.workspace = true
 common-datasource.workspace = true
 common-error.workspace = true
 common-frontend.workspace = true
+common-function.workspace = true
 common-grpc.workspace = true
 common-macro.workspace = true
 common-meta.workspace = true
