+++ src/operator/src/lib.rs
@@ -18,6 +18,7 @@
 pub mod delete;
 pub mod error;
 pub mod expr_factory;
+pub mod flow;
 pub mod insert;
 pub mod metrics;
 pub mod procedure;
